Move latency model converters to execution crate
- Move the latency model converters into the execution crate
- Reuse them from the backtest Python config and engine bindings
- Drop the unused `py` parameter from the extraction converter
